deprecating old sample
